# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Restaurantes",
        "description": "Listado de endpoints de la api",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/"
        }
    ],
    "paths": {
...
        "/api/dishes": {
            "get": {
                "tags": [
                    "Platos"
                ],
                "summary": "Listar todos los platos",
                "description": "Recupera una lista de todos los platos disponibles.",
                "operationId": "ddf841d9ba35d917bd22a504c13b2f72",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "dishes": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "category_name": {
                                                        "type": "string",
                                                        "example": "Categoria"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Plato"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descripción del plato"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "10.50"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Platos"
                ],
                "summary": "Crear un nuevo plato",
                "description": "Crea un nuevo plato en el sistema.",
                "operationId": "3fc0252a7de74dac3747d26739c844cb",
                "requestBody": {
                    "description": "Datos del nuevo plato",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Nombre del nuevo plato"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del nuevo plato"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "12.99"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Insumo guardado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/dish/{id}": {
            "get": {
                "tags": [
                    "Platos"
                ],
                "summary": "Obtener un plato por su ID",
                "description": "Recupera la información de un plato específico por su ID.",
                "operationId": "6e9b0c155f8f8739d4bf6abe56f02e23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del plato a recuperar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "dish": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Platos"
                ],
                "summary": "Actualizar un plato existente",
                "description": "Actualiza la información de un plato existente en el sistema.",
                "operationId": "429651375416801d92426feb5838bd41",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del plato a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del plato a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo nombre del plato"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nueva descripción del plato"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "15.99"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Insumo editado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Platos"
                ],
                "summary": "Eliminar un plato existente",
                "description": "Elimina un plato existente del sistema.",
                "operationId": "21511d45f72d655fc0fe476e88160584",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del plato a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto eliminado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}